+++ lib/theme/app_theme.dart
@@ -1,0 +1,4 @@
+library app_theme;
+
+export 'material_theme.dart';
+export 'text_theme.dart';